+++ src/pyoma2/support/utils/typing.py
@@ -6,18 +6,20 @@ https://github.com/pydantic/pydantic/issues/7017
 
 from __future__ import annotations
 
+import typing
+
 import numpy as np
 from pydantic import BeforeValidator, PlainSerializer
 from typing_extensions import Annotated
 
 
-def nd_array_custom_before_validator(x):
-    # custome before validation logic
+def nd_array_custom_before_validator(x: typing.Any) -> np.ndarray:
+    """Custom before validation logic for numpy arrays."""
     return x
 
 
-def nd_array_custom_serializer(x):
-    # custome serialization logic
+def nd_array_custom_serializer(x: np.ndarray) -> str:
+    """Custom serialization logic for numpy arrays."""
     return str(x)